Move light Gizmo Drawing to core
