+++ com.unity.render-pipelines.core/CoreRP/Editor/Lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a20277bec4b2cc14dbfbc6e401f1f8f1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
